chore(rename): llm_router → llmpop (package, imports, docs, tests)
