+++ .gitignore
@@ -188,4 +188,5 @@ dist/
 .cache/
 running_checks.ipynb
 notebook_for_bash_commands.ipynb
+make_release_notebook.ipynb
 =*
