+++ examples/quick_run_llm_router.py
@@ -1,10 +1,10 @@
-# Quick run of LLM-Router:  
+# Quick run of LLMPop:  
 # -------
 # Quickest on Colab (In Edit, pick the free `T4 GPU`), just copy and paste code.
 # Make sure to install:
-# pip -q install git+https://github.com/LiorGazit/llm_router.git 
+# pip -q install git+https://github.com/LiorGazit/llmpop.git 
 
-from llm_router import init_llm, start_resource_monitoring
+from llmpop import init_llm, start_resource_monitoring
 from langchain_core.prompts import ChatPromptTemplate
 
 # Spinning up OpenAI's free GPT-OSS-20B, give it a few minutes, it's worth it.
