+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "llm_router"                     
+name = "llmpop"                     
 version = "0.2.1"
 description = "The Python library that lets you spin up any LLM with a single function. Route between local and remote LLMs with a unified interface."
 readme = "README.md"
@@ -26,10 +26,10 @@ dependencies = [
 
 [tool.setuptools.packages.find]
 where = ["src"]
-include = ["llm_router*"]
+include = ["llmpop*"]
 
 [tool.setuptools.package-data]
-llm_router = ["py.typed"]
+llmpop = ["py.typed"]
 
 [project.optional-dependencies]
 openai = ["langchain-openai>=0.2.0"]
@@ -43,8 +43,8 @@ dev = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/LiorGazit/llm_router"
-Issues = "https://github.com/LiorGazit/llm_router/issues"
+Homepage = "https://github.com/LiorGazit/llmpop"
+Issues = "https://github.com/LiorGazit/llmpop/issues"
 
 [tool.ruff]
 exclude = ["examples/*.ipynb", ".venv", "build", "dist"]
